requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0
django-environ==0.9.0
environ==1.0
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyotp==2.8.0